Removed some vehicles from no wheel tuning list
For some reason tyrus, tampa and trophytruck were included in a list of vehicles that don't support wheel mods, preventing selection in the menyoo. Commented out to enable re-instatement if there is good reason to have them included. I would like to get a better solution that actually reads the wheel list for a given vehicle to dynamically generate this list.